feat: add AlladdEventListenerCards
